accounts/requirements-dev.txt
aioredis==1.3.1
asgi-cors==0.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.4
dj-static==0.0.6
Django==3.1
django-allauth==0.44.0
django-appconf==1.0.4
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
hashids==1.3.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
oauthlib==3.1.0
payjp==0.1.0
pilkit==2.0
Pillow==8.1.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pyOpenSSL==20.0.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
stripe==2.57.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
whitenoise==5.2.0
zope.interface==5.4.0
